requirements.txt
requests
pyyaml
pymysql
psycopg2-binary
